Passes all tests with Hadoop's I/O streams and customer data.

+++ test/org/anarres/lzo/Lzo1xTest.java
@@ -26,7 +26,7 @@ public class Lzo1xTest {
 
 		byte[] compressed = new byte[orig.length * 2];
 		lzo_uintp compressed_length = new lzo_uintp(compressed.length);
-		int compressed_code = LzoCompressor1x.lzo1x_1_compress(orig, 0, orig.length, compressed, 0, compressed_length, new short[1 << 14]);
+		int compressed_code = LzoCompressor1x.lzo1x_1_compress(orig, 0, orig.length, compressed, 0, compressed_length, new int[1 << 14]);
 
 		System.out.println("Compressed: " + LzoErrors.toString(compressed_code));
 		// System.out.println("Compressed: " + Arrays.toString(Arrays.copyOf(compressed, compressed_length.value)));
